pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "eurelis_kb_framework"
version = "0.0.4a1"
authors = [
  { name="Jérôme DIAZ", email="j.diaz@eurelis.com" },
  { name="Vincent LAMBERT", email="v.lambert@eurelis.com" }
]
description = "Framework to build and manage knowledge base. Based on Langchain under the hood."
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "python-dotenv==1.0.0",
    "langchain==0.0.306",
    "sqlalchemy==2.0.21",
    "rich==13.4.2",
    "click==8.1.7",
    "numpy==1.26.0"
]

[project.optional-dependencies]
huggingface = [
    "sentence_transformers==2.2.2",
    "transformers==4.34.1"
]
openai = [
    "openai==0.28.1"
]
solr = [
    "eurelis-langchain-solr-vectorstore==0.0.1"
]
chroma = [
    "chromadb==0.4.13"
]
sitemap = [
    "lxml==4.9.3",
    "beautifulsoup4==4.12.2"
]
chatbot = [
    "gradio==3.47.1"
]
pdf = [
    "pypdf==3.16.4",
    "cryptography==41.0.7"
]
selfcheck = [
    "lark==1.1.8"
]
mongodb = [
    "pymongo==4.6.1"
]

[project.urls]
"Homepage" = "https://github.com/Eurelis/Eurelis-KB-Framework"
"Bug Tracker" = "https://github.com/Eurelis/Eurelis-KB-Framework/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[project.scripts]
kbf = "eurelis_kb_framework.main:main_cli"

